pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-serving-perf-lab"
version = "0.1.0"
description = "Hiring-signal-first LLM serving performance lab."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
  "pyarrow>=18.0.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.0.0",
  "mypy>=1.10.0",
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]

[project.scripts]
lsp = "lsp.cli.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["lsp*"]

[tool.black]
line-length = 100
target-version = ["py312"]
extend-exclude = '''
/(\
    \.uv-cache
  | \.venv
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "gpu: requires GPU resources",
  "network: requires network access",
]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["lsp"]

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true
